refactor(declarative): split merge generation into collection_tokens, guards, and merge_tokens
Refactored the declarative merge code by extracting collection token generation, guard generation, and merge token generation
into separate submodules. This modularization improves code organization, separates concerns, and reduces the size of the main declarative mod.rs file.
